Move tx related code to GATx for re-use
