+++ app/src/main/java/com/greenaddress/greenbits/ui/SendFragment.java
@@ -54,8 +54,6 @@ import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.SortedSet;
-import java.util.TreeSet;
 
 import de.schildbach.wallet.ui.ScanActivity;
 
@@ -719,87 +717,6 @@ public class SendFragment extends SubaccountFragment {
         });
     }
 
-    private static double extractRate(final Map feeEstimates, final Integer blockNum) {
-        final Map estimate = (Map) feeEstimates.get(Integer.toString(blockNum));
-        return Double.parseDouble(estimate.get("feerate").toString());
-    }
-
-    // FIXME: Duplicated in TransactionActivity.java
-    // Return the best estimate of the fee rate in satoshi/1000 bytes
-    private static Coin getFeeEstimate(final GaService service, final boolean isInstant) {
-        final Map<String, Object> feeEstimates = service.getFeeEstimates();
-        Double bestInstant = null;
-
-        // Iterate the estimates from shortest to longest confirmation time
-        final SortedSet<Integer> keys = new TreeSet<>();
-        for (final String block : feeEstimates.keySet())
-            keys.add(Integer.parseInt(block));
-
-        for (final Integer blockNum : keys) {
-            if (!isInstant && blockNum < 6)
-                continue; // Non-instant: Use 6 confirmation rate and later only
-
-            double rate = extractRate(feeEstimates, blockNum);
-            if (rate <= 0.0)
-                continue; // No estimate available: Try next confirmation rate
-
-            if (isInstant) {
-                // For instant, increase the rate to increase the likelyhood of confirmation.
-                // We use the lowest value of:
-                // a) 1.1 * the 1st or 2nd block fee rate
-                // b) 2.0 * the first rate later than 2 blocks
-                if (blockNum <= 2) {
-                    if (bestInstant == null)
-                       bestInstant = rate * 1.1; // Save earliest fast confirmation rate
-                    continue; // Continue to find the first non-fast rate
-                } else
-                    rate *= 2.0;
-            }
-
-            if (bestInstant != null && bestInstant < rate)
-                rate = bestInstant; // Use the lowest instant rate found
-
-            return Coin.valueOf((long) (rate * 1000 * 1000 * 100));
-        }
-
-        if (bestInstant != null) {
-            // No non-fast confirmation rate, return the fast confirmation rate
-            return Coin.valueOf((long) (bestInstant * 1000 * 1000 * 100));
-        }
-
-        // We don't have a usable fee rate estimate, use a default.
-        if (GaService.IS_ELEMENTS)
-            return Coin.valueOf(1);
-        if (Network.NETWORK == MainNetParams.get())
-            return Coin.valueOf((isInstant ? 200 : 120) * 1000);
-        return Coin.valueOf((isInstant ? 75 : 60) * 1000);
-    }
-
-
-    private Coin addUtxo(final Transaction tx,
-                         final List<JSONMap> utxos, final List<JSONMap> used) {
-        return addUtxo(tx, utxos, used, null, null, null, null);
-    }
-
-    private Coin addUtxo(final Transaction tx,
-                         final List<JSONMap> utxos, final List<JSONMap> used,
-                         final List<Long> inValues, final List<byte[]> inAssetIds,
-                         final List<byte[]> inAbfs, final List<byte[]> inVbfs) {
-        final JSONMap utxo = utxos.get(0);
-        final GaService service = getGAService();
-        utxos.remove(0);
-        if (utxo.getBool("confidential")) {
-            inAssetIds.add(utxo.getBytes("assetId"));
-            inAbfs.add(utxo.getBytes("abf"));
-            inVbfs.add(utxo.getBytes("vbf"));
-        }
-        used.add(utxo);
-        GATx.addInput(service, tx, utxo);
-        if (inValues != null)
-            inValues.add(utxo.getLong("value"));
-        return utxo.getCoin("value");
-    }
-
     private int createRawTransaction(final List<JSONMap> utxos, final String recipient,
                                      final Coin amount, final JSONMap privateData,
                                      final boolean sendAll) {
@@ -810,7 +727,7 @@ public class SendFragment extends SubaccountFragment {
         final GaActivity gaActivity = getGaActivity();
         final GaService service = getGAService();
         final List<JSONMap> used = new ArrayList<>();
-        final Coin feeRate = getFeeEstimate(service, privateData.getBool("instant"));
+        final Coin feeRate = GATx.getFeeEstimate(service, privateData.getBool("instant"));
 
         final Transaction tx = new Transaction(Network.NETWORK);
         tx.addOutput(amount, Address.fromBase58(Network.NETWORK, recipient));
@@ -822,7 +739,7 @@ public class SendFragment extends SubaccountFragment {
 
         // First add inputs until we cover the amount to send
         while ((sendAll || total.isLessThan(amount)) && !utxos.isEmpty())
-            total = total.add(addUtxo(tx, utxos, used));
+            total = total.add(GATx.addUtxo(service, tx, utxos, used));
 
         // Then add inputs until we cover amount + fee/change
         while (true) {
@@ -835,7 +752,7 @@ public class SendFragment extends SubaccountFragment {
                 if (utxos.isEmpty())
                     return R.string.insufficientFundsText; // None left, fail
 
-                total = total.add(addUtxo(tx, utxos, used));
+                total = total.add(GATx.addUtxo(service, tx, utxos, used));
                 continue;
             }
 
@@ -944,7 +861,7 @@ public class SendFragment extends SubaccountFragment {
         final GaActivity gaActivity = getGaActivity();
 
         final List<JSONMap> used = new ArrayList<>();
-        final Coin feeRate = getFeeEstimate(service, privateData.getBool("instant"));
+        final Coin feeRate = GATx.getFeeEstimate(service, privateData.getBool("instant"));
 
         final ElementsTransaction tx = new ElementsTransaction(Network.NETWORK);
 
@@ -967,7 +884,7 @@ public class SendFragment extends SubaccountFragment {
 
         // First add inputs until we cover the amount to send
         while (total.isLessThan(amount) && !utxos.isEmpty())
-            total = total.add(addUtxo(tx, utxos, used, inValues, inAssetIds, inAbfs, inVbfs));
+            total = total.add(GATx.addUtxo(service, tx, utxos, used, inValues, inAssetIds, inAbfs, inVbfs));
 
         // Then add inputs until we cover amount + fee/change
         while (true) {
@@ -979,7 +896,7 @@ public class SendFragment extends SubaccountFragment {
                 // Need more inputs to cover amount + fee/change
                 if (utxos.isEmpty())
                     return R.string.insufficientFundsText; // None left, fail
-                total = total.add(addUtxo(tx, utxos, used, inValues, inAssetIds, inAbfs, inVbfs));
+                total = total.add(GATx.addUtxo(service, tx, utxos, used, inValues, inAssetIds, inAbfs, inVbfs));
                 continue;
             }
